made restockable inputs work properly for processing tasks

+++ src/main/java/com/raoulvdberge/refinedstorage/api/storage/externalstorage/IStorageExternal.java
@@ -24,5 +24,5 @@ public interface IStorageExternal<T> extends IStorage<T> {
     /**
      * @return the capacity of the connected storage
      */
-    int getCapacity();
+    long getCapacity();
 }
